% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_instructor.R
\name{add_instructor}
\alias{add_instructor}
\title{Add an instructor to the \code{instructors} data object}
\usage{
add_instructor(names, email)
}
\arguments{
\item{email}{Character vector. Email as it appears in the the instructors'
IBKR paper trader account.}

\item{name}{Character vector. Instructors' name as it appears in
the instructor's IBKR paper trader account.}
}
\description{
Takes in character vectors containing instructors' names and emails and
identifies them as instructors within the package, meaning that performance
metrics won't be calculated for these accounts.
}
