% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_email_performance_reports.R
\name{create_and_email_performance_reports}
\alias{create_and_email_performance_reports}
\title{Email Performance Reports}
\usage{
create_and_email_performance_reports(
  subject = "2020 FINTECH Trading Competition Update",
  body = "Updated standings attached!",
  attachments = TRUE
)
}
\description{
Email Performance Reports
}
