% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_performance_report_pdf_and_excel_files.R
\name{create_performance_report_pdf_and_excel_files}
\alias{create_performance_report_pdf_and_excel_files}
\title{Create Individual Report PDF & Excel Files}
\usage{
create_performance_report_pdf_and_excel_files(excel = FALSE)
}
\description{
Create Individual Report PDF & Excel Files
}
