% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrr.R
\name{gmrr}
\alias{gmrr}
\title{gmrr}
\usage{
gmrr(returns_xts)
}
\arguments{
\item{returns_xts}{An xts object whose values are period-over-period returns
observed for the assets specified by the identifiers in the column names.
Returns are NOT assumed to be in 'percent form': i.e., make sure that in
whatever xts is passed as \code{returns_xts}, a return of \emph{12}\% is
represented as \emph{0.12}.}
}
\value{
A numeric vector whose values are the geometric means of the returns
in \code{returns_xts} and whose names are the identifiers for which each
geometric mean return was calculated.
}
\description{
Calculate geometric means of an xts series of returns.
}
