% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_benchmark_plot.R
\name{make_benchmark_plot}
\alias{make_benchmark_plot}
\title{Create a Benchmark Plot}
\usage{
make_benchmark_plot(portfolio_returns, benchmark_asset, x_lab)
}
\description{
Create a Benchmark Plot
}
