% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_all_data.R
\name{refresh_all_data}
\alias{refresh_all_data}
\title{Update the student.statements.rda and benchmark.rda package data objects}
\usage{
refresh_all_data()
}
\arguments{
\item{filename}{Name of the xml file containing student statement data.

Refreshes benchmark.rda, gspc.rda, performance_reports.rda, rfr.rda, and
student_reports.rda. Stores as package data objects.}
}
\description{
Imports the results of a Flex Query run within IBKR and downloaded as an xml
file, fetches the relevant data, and saves a \link[tibble]{tibble} containing
the student statement information needed to calculate standings and
performance metrics as the package data object \emph{student_reports.rda}. Same
for the benchmark.
}
