% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_base_data.R
\name{refresh_base_data}
\alias{refresh_base_data}
\title{Updates all base data that the app needs to generate the site.}
\usage{
refresh_base_data()
}
\description{
Updates all base data that the app needs to generate the site.
}
