% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registry}
\alias{registry}
\title{Registry}
\format{
A data frame
}
\usage{
registry
}
\description{
Dataframe containing registered participants by tradername alongside their
status in the onboarding process.
}
\keyword{datasets}
